BUG FIX: Correção de todos arquivos que trabalhavam quantidade em Integer

+++ src/main/java/com/pdv/lalapan/dto/VendaAddItemRequestDTO.java
@@ -1,5 +1,7 @@
 package com.pdv.lalapan.dto;
 
-public record VendaAddItemRequestDTO(Long idProduto, int quantidade) {
+import java.math.BigDecimal;
+
+public record VendaAddItemRequestDTO(Long idProduto, BigDecimal quantidade) {
 
 }
