+++ src/main/java/com/pdv/lalapan/entities/VendaItens.java
@@ -10,21 +10,7 @@ public class VendaItens {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    /*
-    * TODO: CORRIGIR O ATRIBUTO "quantidade".
-    *  1°: substituir toda e qualquer lógica que envolva calculo de subtotal, pois atualmente
-    *      os calculos estão sendo feitos com o atributo "quantidade" em Integer, se uma quantidade
-    *      for lida em decimal (gramas ou kg decimal) o calculo não trabalha precisamente, convertendo
-    *      para um numero inteiro.
-    *
-    * TODO: INSPECIONAR FUNCIONAMENTO.
-    *  2°: inspecionar o funcionamento de cada etapa do ciclo de venda para ter certeza de que o calculo
-     *     está sendo feito da maneira correta.
-     * 3°: Iniciar implementação do ciclo de calculo nos testes de integração e unitários
-     */
-
-    private Integer quantidade;
+    private BigDecimal quantidade;
 
     private BigDecimal precoUnitario;
     private BigDecimal subtotal;
@@ -49,11 +35,11 @@ public class VendaItens {
         this.id = id;
     }
 
-    public Integer getQuantidade() {
+    public BigDecimal getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(Integer quantidade) {
+    public void setQuantidade(BigDecimal quantidade) {
         this.quantidade = quantidade;
     }
 
@@ -90,6 +76,6 @@ public class VendaItens {
     }
 
     public void calcularSubTotal() {
-        this.subtotal = precoUnitario.multiply(BigDecimal.valueOf(quantidade));
+        this.subtotal = precoUnitario.multiply(quantidade);
     }
 }
